requirements.txt
pytesseract==0.3.13
Pillow==11.1.0
PyMuPDF==1.25.1
